The topup dialog now functions and persists on database level. Fix a bug where the current student's total and unused count is not correctly reset if the operation is cancelled.
jQuery mobile enhances DOM and the enhanced method cannot be used before
an element is enhanced. This creates some issue when sync slider widgets
when one is enhanced and the other is not and later both of them are
enhanced. Can used the data method to check if an element is enhanced,
e.g. data("moible-slider").
Will look into merge the two stamps page by using two divs and animate
the transition manually later. But it is not current priority as the
changes may not be very straightforward and the current code achieves a
similar effect albeit with ulgy code.