Re-work on the message recipient selection UI which works better and is more controlable than the last attempt.
The UI is stateless which means it does not remember the previous
selection regardless whether the message is sent or discarded.
The UI implementation indicates that the message can be sent to multiple
lessons, which is different from current sever side message system (only
single recipient lesson is allowed). The News tab is also implemented with
single recipient lesson in mind. They both need to be re-worked on.